% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_correlation.R
\name{s_corr}
\alias{s_corr}
\title{Sample Correlation}
\usage{
s_corr(experiment)
}
\arguments{
\item{experiment}{A two dimensional dataframe of size NxP}
}
\value{
The RxR correlation matrix for the experimental data.
}
\description{
Calculates the sample correlation matrix for a given NxP
multivariate dataframe, where N is the number of trials in the
experiment and P is the number of variables.
}
\examples{
# generate 10x3 dataframe of random values for testing
fake_data = data.frame(replicate(3,sample(100,10,rep=TRUE))) / 100

# run fake data through the custom sample correlation function
test_result = s_corr(fake_data);

}
