% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_biased_covariance.R
\name{s_cov}
\alias{s_cov}
\title{Biased Sample Covariance}
\usage{
s_cov(experiment)
}
\arguments{
\item{experiment}{A two dimensional dataframe of size NxP}
}
\value{
The RxR covariance matrix for the experimental data.
}
\description{
Calculates the biased sample covariance matrix for a given NxP
multivariate dataframe, where N is the number of trials in the
experiment and P is the number of variables.
}
\examples{
# generate 10x3 dataframe of random values for testing
fake_data = data.frame(replicate(3,sample(100,10,rep=TRUE))) / 100

# run fake data through the custom sample covariance function
test_result = s_cov(fake_data);

}
