% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{s_mean}
\alias{s_mean}
\title{Sample Mean}
\usage{
s_mean(experiment)
}
\arguments{
\item{experiment}{A two dimensional dataframe of size NxP}
}
\value{
The 1xR vector of sample means for the experimental data.
}
\description{
Calculates the sample mean for a given NxP multivariate dataframe,
where N is the number of trials in the experiment and P is the number
of variables.
}
\examples{
# generate 10x3 dataframe of random values for testing
fake_data = data.frame(replicate(3,sample(100,10,rep=TRUE))) / 100

# run fake data through the custom sample mean function
test_result = s_mean(fake_data);

}
